requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.3
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-ratelimit==4.1.0
django-redis==6.0.0
django-sendgrid-v5==1.3.0
djangorestframework==3.16.1
idna==3.11
kombu==5.5.4
Markdown==3.9
MarkupSafe==3.0.3
packaging==25.0
prompt_toolkit==3.0.52
pycparser==2.23
whitenoise
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
redis==7.1.0
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
